requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.8
django-tinymce==3.0.2
django-tinymce4-lite==1.7.4
djangorestframework==3.11.0
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jsmin==2.2.2
lxml==4.5.2
oauthlib==3.1.0
Pillow==7.2.0
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.1.0
